Add Lerp trait and implementation.
This trait will be a fundamental building block of the timeline types. Any value can be included in a timeline as long as it supports this trait.

There is already a `lerp` crate with a nice derive macro, but it doesn't have quite the right API for this application. It is focused on float types and structs containing floats; the goal here is to automatically support all numeric primitives (i.e. including integer types) and we aren't very concerned with bounds, etc.