Add high-level timeline types and sub-timeline impl.
Sub-timeline design should be more CPU-optimal (although it may take slightly more memory) and enables a substantial amount of the functionality to be written and tested without having to deal with the proc macros. The tests also provide a pretty good template for what the macro output should look like, although some aspects (e.g. looping) are not incorporated.