Add a time-scale abstraction.
This will help to logically separate the "keyframes" oriented view represented by a `SubTimeline` and specified in the generated builders from the "real time" behavior of an actual animator. Keyframes are forward-only and non-repeating, and must be the latter to avoid infinitely-sized data; we instead track a "transport position" which is bounded to the duration and can run in either direction.

Importantly, this is a stateless type, and does not track any actual position at any given instant. Elapsed time will be logically part of the animator state and may be implemented differently depending on the context (bevy, iced, something else).