First working graphical example.
Uses nannou for quick set up. The downside is that this pulls in quite a lot of dependencies and takes a long time to build, but can be mitigated by switching to LLD, such as in the [Bevy setup](https://bevyengine.org/learn/book/getting-started/setup/#enable-fast-compiles-optional). The upside is that it dramatically reduces the amount of pointless boilerplate compared to e.g. a pure wgpu example, so that the examples can really focus on the animation code and its simplicity.