Rework Timeline API to be update-based rather than query-based.
While this is slightly impure, it solves a few problems that are otherwise difficult to solve. First, it allows properties that are animatable but not in the specific timeline to be ignored entirely, which should always have been the expected behavior, and is fixed in this commit. Second, it's very likely to be more performant, since it doesn't require the constant creation of new value structs.

This ends up actually simplifying the code generation because it turns out we don't need the generated `AnimatorValues` struct at all. The only remaining usage was to get the sub-timeline defaults for when 0% keyframes aren't specified, but the `Default` implementation for that struct was derived anyway, i.e. using the defaults for all fields, so we get the same result by supplying `Default::default()` for the individual fields.