Add timeline merging.
Includes tests and example. Merging is important when animations involve loops or transitions that have different durations, different repeat or reverse behavior, or major differences in easing that could look jerky in practice. When a merged timeline is used, the individual timelines all keep their individual behavior and "overlaid" on top of each other in time.