Allow timelines to change their start values.
Transitioning between animations always involves the new animation starting where the previous animation left off. More advanced blending is possible, but this is the very basic form.

Because this is a mutation, most of the types have been updated to support and/or require `Clone`. It is not unlikely for a caller to want to preserve the original timeline, or perhaps even store it in some app-level resource and reuse it across many widgets. However, once they _are_ cloned and stored per widget then it is no longer necessary to keep cloning them, which is why the explicit option to clone-then-mutate is better than a method that always clones and returns a new instance.

An immutable wrapper that can reuse most of the (also assumed to be immutable) timeline data would be even better, but is complicated to maintain with Rust's borrowing rules.

No example for this one as it's just a building block for state-triggered animations.