Adds state animator and builder API.
The resulting syntax is almost identical to the original proposal. It is still a bit noisy compared to the TBD derive macro, but the use of `TimelineOrBuilder` helps cut most of it - we only need the one `build` call at the end.

In general, the expectation is that anything which wants to store an animator should be a generic type with `<A: StateAnimator>`. We can't really hide the `MappedTimelineAnimator` completely without causing a lot of difficulty with macros and forcing downstream users to use `dyn Box`, but it's not intended to be used directly.

Still needs the macro (probably mbe) to clean it up fully.