Add an iced example.
Fixes a few bugs in the animator macro that weren't covered by tests, mainly incorrect parsing of the easing function, failing to actually configure the easing function once recognized, and an implicit dependency on a syn feature that needed to be made explicit.

Adds a few nice-to-have features uncovered by the example, including a current-state getter and a type alias to avoid the need for either boxing or long generics.

Have not figured out a way to avoid requiring the app to declare an explicit dependency on `enum-map` in order to use the aliased `State` derive. May be inherent in the enum-map implementation itself, which is written with absolute crate paths.