Support pseudo pause/resume based on state.
If a particular state is not present at all in an animator (as opposed to being associated with an empty timeline), then that state can act as a "pause", and switching back to a state that was previously animating will cause the previous animation to _resume_ instead of blending or starting over.

Because of blending, the difference can be subtle, but it is most noticeable if the "paused" animation is long, looping, repeating, or all of the above. For example, if we "pause" at 8 seconds out of a 10-second animation, then with this change it will take 2 seconds after "resume" to complete or loop, instead of another full 10 seconds at a seemingly slower speed.