Refactor `TimeScale` to provide repeat/reverse info.
In order for an animator to revert start values on repetition or reverse, it is going to need this info every time it advances. Convert the original `Result` into a regular enum, whose `Active` variant holds the original normalized time and also the required information about whether the time is on a repeat or reverse pass.

Also, the not-started and ended results being considered normal results, as opposed to an `Err`, will make it simpler later on to have these states handled and make the animator stop running or perform follow-up actions (like potentially stopping the actual animation timer) when these states are reached, although that's not a primary goal.