Restrict timeline blending to first cycle.
When transitioning states, the values of the previous state should not be "sticky". This difference isn't apparent for single-cycle animations, but when the animations reverse or repeat, it can be very obvious when the new animation is repeating from or reversing to a value that may not even be in the correct _direction_.

By piping the `TimeScaleLoopState` through a couple of methods, we can keep this behavior mostly stateless, i.e. not have to change mutability of any timeline-related types or methods. The `Timeline::update` method receives a real time and therefore can figure out everything it needs to decide whether to use blended or original values. It does make the `SubTimeline` API somewhat unwieldy, but that is just an internal helper API anyway.