Add an Iced widget example.
This is by far the most complex example so far and hopes to set out some useful design patterns for integrating with Iced. Self-animated widgets are really the way to go, because otherwise responding to changes in state requires constant clumsy message-passing between the widgets and the view/app and publicly exposing a lot of private state through either its public interface or its stylesheet.

Being able to attach arbitrary animators to widgets would be even better (like how web animations are defined in the style sheet), but to support that would probably require changes to Iced itself. For now, this is a pretty decent halfway point.