Partial Bevy example.
Includes a plugin that combines animators with the `bevy_mod_picking`. With some refactoring and fine-tuning, it might make sense to release this as an official extension, since it's not especially complex but still potentially nonintuitive to set up. The best approach would be to separate the animation and picking+animation into separate plugins.

The current example is just a few nav buttons; the complete example will also have some animated content in the middle.