Add a `timeline` macro and update some examples to use it.
It's obvious this macro should exist because a superset of the syntax is already supported at the animator level, and non-toy apps are supposed to be able to use individual timelines as well (as the Bevy example clearly demonstrates).

Bevy example itself is _not_ updated yet because it relies on keyframe-specific easing and the syntax doesn't support that yet. The merged timeline example is where the most significant readability improvement can be seen.