Separate registration of Bevy plugin from key registration.
It's not necessarily ergonomic to have to do both at the same time. This is especially true in Bevy apps where one module might handle the core animation, and install the plugins for animating certain components, while another module (or several) might use its own keys to control those animations.

Case in point would be a `Transform`, which is used on almost every "visible" entity, but will almost never be driven by a state that's common to all entities.