Add a real `Animate` trait that holds the builder/helper methods.
The public API does not really change, except that the `timeline` and `keyframe` methods are now trait methods instead if `impl` methods on the animated struct. The main purpose of this change is to provide some additional tooling support when dealing with generic timeline types _before_ any builders are created (e.g. creating "tweens").