Allow animators to be disabled.
Having them running all the time leads to some difficulties controlling timing in real-world scenarios; e.g. we want to run some animation, then hide the animated entity, and then reset the state of the animated component so it looks right when re-shown. If the timing isn't extremely tight then it's OK to just reset the animator and let it control the state. But if timing is very tight, and there's a risk that the post-animation state could be visible before the animator has a chance to reset it, then disabling the animator is necessary so that it's possible to reset the animated properties without having them animate again.